WEBSITE/artist-hub/backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
google-cloud-storage
